refactor: 클래스 의존관계 설정 변경
StationLine 클래스 제거 후 Section 클래스 생성
Line에서 StationLine 의존관계 제거
Station에서 StationLine  의존관계 제거
Staions라는 일급컬렉션 생성
