+++ src/main/java/subway/domain/line/Line.java
@@ -1,13 +1,7 @@
 package subway.domain.line;
 
-import subway.domain.StationLine;
-
-import java.util.ArrayList;
-import java.util.List;
-
 public class Line {
     private String name;
-    private List<StationLine> stationLines = new ArrayList<>();
 
     private Line(String name) {
         this.name = name;
