+++ src/main/java/subway/domain/station/Station.java
@@ -1,13 +1,7 @@
 package subway.domain.station;
 
-import subway.domain.StationLine;
-
-import java.util.ArrayList;
-import java.util.List;
-
 public class Station {
     private String name;
-    private List<StationLine> stationLines = new ArrayList<>();
 
     private Station(String name) {
         this.name = name;
