+++ src/main/java/subway/domain/station/Stations.java
@@ -1,0 +1,16 @@
+package subway.domain.station;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Stations {
+    private final List<Station> stations;
+
+    private Stations(List<Station> stations) {
+        this.stations = new ArrayList<>(stations);
+    }
+
+    public static Stations of(List<Station> stations) {
+        return new Stations(stations);
+    }
+}
